fix: remove types in compat
